+++ code/qcommon/files.c
@@ -481,6 +481,7 @@ FS_FileForHandle(fileHandle_t f)
   {
     Com_Error(ERR_DROP, "FS_FileForHandle: can't get FILE on zip file");
   }
+
   if (!fsh[f].handleFiles.file.o)
   {
     Com_Error(ERR_DROP, "FS_FileForHandle: NULL");